+++ src/VisuManager.ts
@@ -125,7 +125,7 @@ export class VisuManager {
     const visu = new WebviewVisu(
       "meshViewer",
       testDir,
-      "resources/visu_vtk/index.html",
+      "resources/visu_vtk/dist/index.html",
       fileContexts,
       objUris.map((uri) => path.basename(uri.fsPath)),
       undefined,